[#410] Fix misprint in the 'double.MinValue' definition as expression `double.MinValue + 0` causes error `Number.MIN_VALUE.add is not a function`

+++ Bridge/System/Double.cs
@@ -14,7 +14,7 @@ namespace System
 		[Name("MAX_VALUE")]
 		public static readonly double MaxValue = 0;
         [Name("MIN_VALUE")]
-        public static readonly decimal MinValue = 0;
+        public static readonly double MinValue = 0;
 
 		[InlineConst]
 		public const double Epsilon = 4.94065645841247E-324;
